Refactored Handlers and Routes.
